+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Login from './components/login';
+import Login from './components/Login/Login';
 
 const App = () => {
   return (
